Uploaded Original Docs
Uploaded the Jupyter Notebook files containing the crucial code elements of the project. This includes a model training notebook, and a working system notebook which each represent key elements of the final product while the image segmentation and smaller field dimension notebooks contain algorithms that are employed in the final system but are non-essential.